+++ api/src/v2/routes/seasons/get.integration.test.ts
@@ -1,4 +1,4 @@
-import { getAllSeasons } from "./getAll";
+import { getAllSeasons } from "./get";
 import express from "express";
 import request from "supertest";
 import { useEphemeralDb } from "../../../test/with-ephemeral-db";
