+++ api/src/v2/routes/seasons/get.ts
